+++ Telegram/SourceFiles/ayu/ui/context_menu/context_menu.h
@@ -28,6 +28,10 @@ void AddJumpToBeginningAction(PeerData *peerData,
 void AddOpenChannelAction(PeerData *peerData,
 							  not_null<Window::SessionController*> sessionController,
 							  const Window::PeerMenuCallback &addCallback);
+void AddDeleteOwnMessagesAction(PeerData *peerData,
+								Data::ForumTopic* topic,
+								not_null<Window::SessionController*> sessionController,
+								const Window::PeerMenuCallback &addCallback);
 
 void AddHistoryAction(not_null<Ui::PopupMenu*> menu, HistoryItem *item);
 void AddHideMessageAction(not_null<Ui::PopupMenu*> menu, HistoryItem *item);
